Decoupled Storage from CSP. Corrected definition of Tr (not transmission but working tempeature of fluid)
